Client - remove duration from equipments list
